% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcOR.R
\name{CalcOR}
\alias{CalcOR}
\title{CalcOR Function}
\usage{
CalcOR(vec.abcd, a = NULL, b = NULL, cc = NULL, d = NULL, vect = T)
}
\description{
Calculate Odds ratio from data
}
\examples{
CalcOR(c(10,20,20,50))

}
\keyword{CalcOR}
