% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatIndex.R
\name{CatIndex}
\alias{CatIndex}
\title{CatIndex Function}
\usage{
CatIndex(index.var, index.char = NULL)
}
\arguments{
\item{index.var}{index.char}
}
\description{
Useful for simulations: CatIndex print on screen the infex index.var and its value
}
\examples{
i = 7
CatIndex(i)
CatIndex(i,"index i")
}
\keyword{simulations}
