% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CItrans.R
\name{CItrans}
\alias{CItrans}
\title{CItrans Function}
\usage{
CItrans(coef, se, Ftrans, cover = 0.95)
}
\arguments{
\item{coef}{se Ftrans cover}
}
\description{
CItrans takes coef estimate, its estimated se, calculates CI according to normal asusmption, and then transormate the ci according to the function Ftrans
}
\examples{

}
