% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverCInorm.R
\name{CoverCInorm}
\alias{CoverCInorm}
\title{CoverCInorm Function}
\usage{
CoverCInorm(est, se, truth, cover.prob = 0.95)
}
\description{
Given an estimate and standard error, returns "T" if the CI covers the true value
}
\examples{
CoverCInorm(est = 0.4, se = 0.1, truth = 0.5)

}
\keyword{CalcOR}
