% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{expit Function}
\usage{
expit(x)
}
\arguments{
\item{x}{any number}
}
\description{
This function carries out the usual expit(x)= exp(x)/(1+exp(x)) function
}
\examples{
expit(0)
}
\keyword{logit}

