% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logit Function}
\usage{
logit(p)
}
\arguments{
\item{TRUE.}{}
}
\description{
This function carries out the usual logit(p)=log(p/(1-p)) function
}
\examples{
logit(0.5)
}
\keyword{logit}

