% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargORtoJoint.R
\name{MargORtoJoint}
\alias{MargORtoJoint}
\title{MargORtoJoint Function}
\usage{
MargORtoJoint(p1marg, p2marg, OR)
}
\description{
For a 2X2 talbe, this function takes two marginal probabilities and the odds ratio
and returns 4 cell probabilities
}
\examples{
MargORtoJoint(0.4,0.7,2)
}
\keyword{MargORtoJoint}
